+++ server/controllers/department/parts/get_parts.js
@@ -6,7 +6,7 @@ const GetPartsForTask = async (req, res) => {
         const { rows } = await pool.query(
             `SELECT id, unique_id, ticket_number, part_name, part_desc, 
                     seal_number, part_quantity, parts_status, created_at, 
-                    compensation, created_by, updated_at 
+                    compensation, part_issued, created_by, updated_at, credit_req_number
              FROM parts_for_tasks 
              WHERE task_row_id = $1 
              ORDER BY created_at DESC`,